bot/requirements.txt
python-telegram-bot~=20.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
supabase~=2.10.0
dotenv~=0.9.9
aiohttp
asyncio